Okay we are slowly trying to implement labelled number types
i8, i16, etc.  Just need to massage them into the lexer and parser.

+++ src/lib.rs
@@ -52,7 +52,7 @@ impl From<VarSym> for usize {
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum TypeDef {
     /// Signed integer with the given number of bytes
-    SInt(usize),
+    SInt(u8),
     /// An integer of unknown size, from an integer literal
     UnknownInt,
     Bool,
